# openapi/executor.v1.yaml
openapi: 3.1.0
info:
  title: Polymarket Execution Engine API
  version: 1.0.0-draft
  description: >
    Public adapter-facing API for a standalone deterministic Polymarket execution engine.
    This API intentionally does not expose private keys, CLOB API secrets, raw signed payloads,
    internal signed-order envelopes, or direct database writes.
servers:
  - url: http://localhost:8080
security:
  - ServiceToken: []
paths:
...
  /v1/intents/normalize:
    post:
      operationId: normalizeIntent
      security:
        - ServiceToken: []
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TradeIntent' }
      responses:
        '200':
          description: Normalized intent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NormalizedIntent' }
  /v1/snapshots/capture:
    post:
      operationId: captureSnapshot
      security:
        - ServiceToken: []
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NormalizedIntent' }
      responses:
        '200':
          description: Feasibility snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeasibilitySnapshot' }
  /v1/decisions/evaluate:
    post:
      operationId: evaluateDecision
      security:
        - ServiceToken: []
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DecisionRequest' }
      responses:
        '200':
          description: Constraint decision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConstraintDecision' }
  /v1/plans/compile:
    post:
      operationId: compilePlan
      security:
        - ServiceToken: []
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompilePlanRequest' }
      responses:
        '200':
          description: Execution plan summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionPlanSummary' }
  /v1/submissions:
    post:
      operationId: submitPlan
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: header
          name: X-Correlation-Id
          required: false
          schema: { type: string }
          description: Optional executor correlation id propagated into submit lifecycle evidence.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitRequest' }
      responses:
        '202':
          description: Submit accepted into executor-owned state machine; scaffold remains blocked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubmitReceipt' }
  /v1/submissions/{execution_id}:
    get:
      operationId: getSubmission
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: path
          name: execution_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Submit receipt/status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubmitReceipt' }
  /v1/sign-only/lifecycle-events:
    post:
      operationId: recordSignOnlyLifecycleEvent
      security:
        - ServiceToken: []
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignOnlyLifecycleRecord' }
      responses:
        '202':
          description: Sign-only lifecycle event accepted after local state-machine validation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignOnlyLifecycleRecord' }
  /v1/sign-only/standard-constructions:
    post:
      operationId: recordStandardSignOnlyConstruction
      security:
        - ServiceToken: []
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StandardSignOnlyConstructionRequest' }
      responses:
        '202':
          description: Standard SDK sign-only construction recorded without remote side effects
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StandardSignOnlyConstructionReceipt' }
  /v1/sign-only/lifecycle-events/{execution_id}:
    get:
      operationId: listSignOnlyLifecycleEvents
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: path
          name: execution_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: before_event_id
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Sign-only lifecycle records for one execution
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SignOnlyLifecycleRecord' }
  /v1/lifecycle/executions/{execution_id}/events:
    get:
      operationId: listExecutionLifecycleEvents
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: path
          name: execution_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: before_event_id
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Execution lifecycle audit events for one execution
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExecutionLifecycleEvent' }
  /v1/lifecycle/orders/{order_id}/events:
    get:
      operationId: listOrderLifecycleEvents
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: path
          name: order_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: before_event_id
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Local order lifecycle events for one order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OrderLifecycleEventRecord' }
  /v1/runtime/workers:
    get:
      operationId: listRuntimeWorkerStatus
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: query
          name: account_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: before_observed_at
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Recent runtime worker heartbeats and account-scoped fail-closed observations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuntimeWorkerStatusReport' }
  /v1/portfolio/projections:
    post:
      operationId: recordPortfolioProjection
      description: >
        Non-live portfolio projection preparation endpoint. Records local fills,
        positions, open orders, and exposure projection state for later readback
        and risk-limit assessment; it does not sign, submit, cancel, or call CLOB.
      security:
        - ServiceToken: []
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortfolioProjection' }
      responses:
        '202':
          description: Portfolio projection accepted into local executor store
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioProjectionRecordResponse' }
  /v1/portfolio/{account_id}/projection:
    get:
      operationId: getPortfolioProjection
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: path
          name: account_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Last accepted local portfolio projection for an account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioProjection' }
  /v1/portfolio/{account_id}/risk-assessments:
    post:
      operationId: assessPortfolioRisk
      description: >
        Non-live risk-limit assessment against the stored local portfolio
        projection. The endpoint returns a decision only; it does not authorize
        live submit/cancel or perform remote side effects.
      security:
        - ServiceToken: []
        - AdminToken: []
      parameters:
        - in: path
          name: account_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RiskLimits' }
      responses:
        '200':
          description: Risk decision derived from stored local exposure projection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskDecision' }
  /v1/admin/session:
    get:
      operationId: getAdminSession
      security:
        - AdminToken: []
        - AdminReadToken: []
      responses:
        '200':
          description: Authenticated admin subject and authorized capability set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSession' }
  /v1/admin/audit-events:
    get:
      operationId: listAdminAuditEvents
      security:
        - AdminToken: []
        - AdminReadToken: []
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - in: query
          name: before_audit_id
          required: false
          schema: { type: integer, minimum: 1 }
        - in: query
          name: operation
          required: false
          schema: { type: string }
        - in: query
          name: principal_subject
          required: false
          schema: { type: string }
        - in: query
          name: result
          required: false
          schema: { type: string }
        - in: query
          name: correlation_id
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Recent admin audit events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminAuditEvent' }
...
  /v1/admin/kill-switch:
    post:
      operationId: setKillSwitch
      security:
        - AdminToken: []
        - EmergencyOperatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KillSwitchRequest' }
      responses:
        '202':
          description: Kill switch update accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KillSwitchReceipt' }
  /v1/admin/cancel-order:
    post:
      operationId: cancelOrder
      description: >
        Non-live cancel scaffold. Records local audit/lifecycle intent and
        returns a reconcile-required receipt; it does not call Polymarket CLOB
        cancel APIs or claim remote cancellation.
      security:
        - AdminToken: []
        - AdminCancelToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CancelOrderRequest' }
      responses:
        '202':
          description: Non-live cancel request accepted into pending/reconcile-required state machine
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CancelReceipt' }
  /v1/admin/reconcile:
    post:
      operationId: reconcile
      description: >
        Non-live reconcile scaffold. Records local reconcile/audit context and
        optional local order-lifecycle divergence; it does not perform remote
        reconciliation reads or live cancel/submit side effects.
      security:
        - AdminToken: []
        - AdminCancelToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReconcileRequest' }
      responses:
        '202':
          description: Non-live reconcile request accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReconcileReport' }
  /v1/admin/reconcile-order-local:
    post:
      operationId: reconcileOrderLocal
      security:
        - AdminToken: []
        - AdminCancelToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReconcileOrderLocalRequest' }
      responses:
        '202':
          description: Local order lifecycle reconciliation accepted without remote side effects
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReconcileOrderLocalResponse' }
components:
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: Service-scope token for adapter-facing operations.
    AdminToken:
      type: http
      scheme: bearer
      description: Legacy full admin-scope token for audit, cancel/reconcile, and kill-switch operations.
    AdminReadToken:
      type: http
      scheme: bearer
      description: Admin-read token for audit/session inspection only.
    AdminCancelToken:
      type: http
      scheme: bearer
      description: Admin-cancel token for non-live cancel and reconcile operations only.
    EmergencyOperatorToken:
      type: http
      scheme: bearer
      description: Emergency-operator token for kill-switch operations only.
  schemas:
...
    Side:
      type: string
      enum: [BUY, SELL]
    TimeInForce:
      type: string
      enum: [GTC, FOK, GTD, FAK]
    MarketRef:
      type: object
      additionalProperties: false
      required: [condition_id, slug, is_sports]
      properties:
        condition_id: { type: string }
        slug: { type: [string, 'null'] }
        is_sports: { type: boolean }
    QuantityIntent:
      type: object
      additionalProperties: false
      required: [max_notional, max_shares]
      properties:
        max_notional: { anyOf: [ { $ref: '#/components/schemas/PositiveCanonicalDecimal' }, { type: 'null' } ] }
        max_shares: { anyOf: [ { $ref: '#/components/schemas/PositiveCanonicalDecimal' }, { type: 'null' } ] }
      description: Exactly one of max_notional or max_shares must be non-null; enforced by executor.
    TradeIntent:
      type: object
      additionalProperties: false
      required: [client_intent_id, account_id, market, token_id, side, quantity, limit_price, time_in_force, collateral_profile_id]
      properties:
        client_intent_id: { type: string }
        account_id: { type: string }
        market: { $ref: '#/components/schemas/MarketRef' }
        token_id: { type: string }
        side: { $ref: '#/components/schemas/Side' }
        quantity: { $ref: '#/components/schemas/QuantityIntent' }
        limit_price: { $ref: '#/components/schemas/LimitPrice' }
        time_in_force: { $ref: '#/components/schemas/TimeInForce' }
        collateral_profile_id: { type: [string, 'null'] }
...
    NormalizedIntent:
      type: object
      additionalProperties: false
      required: [normalized_intent_id, intent_hash, account_id, market, token_id, side, quantity_bound, limit_price, time_in_force, collateral_profile_id]
      properties:
        normalized_intent_id: { type: string }
        intent_hash: { type: string }
        account_id: { type: string }
        market: { $ref: '#/components/schemas/MarketRef' }
        token_id: { type: string }
        side: { $ref: '#/components/schemas/Side' }
        quantity_bound: { $ref: '#/components/schemas/QuantityBound' }
        limit_price: { $ref: '#/components/schemas/LimitPrice' }
        time_in_force: { $ref: '#/components/schemas/TimeInForce' }
        collateral_profile_id: { type: [string, 'null'] }
...
    FeasibilitySnapshot:
      type: object
      additionalProperties: false
      required: [snapshot_id, snapshot_hash, normalized_intent_id, runtime_state, captured_at]
      properties:
        snapshot_id: { type: string }
        snapshot_hash: { type: string }
        normalized_intent_id: { type: string }
        runtime_state: { $ref: '#/components/schemas/RuntimeStateSummary' }
        captured_at: { type: string, format: date-time }
    DecisionRequest:
      type: object
      additionalProperties: false
      required: [normalized_intent_id, snapshot_id]
      properties:
        normalized_intent_id: { type: string }
        snapshot_id: { type: string }
      description: Server-authoritative v0.14 request; executor loads both objects from store and verifies binding.
    ConstraintDecision:
      type: object
      additionalProperties: false
      required: [decision_id, decision_hash, status, reasons]
      properties:
        decision_id: { type: string }
        decision_hash: { type: string }
        status: { type: string, enum: [ALLOW, BLOCK, CLOSE_ONLY, DEGRADED] }
        reasons:
          type: array
          items: { type: string }
...
    CompilePlanRequest:
      type: object
      additionalProperties: false
      required: [normalized_intent_id, snapshot_id, decision_id, approval]
      properties:
        normalized_intent_id: { type: string }
        snapshot_id: { type: string }
        decision_id: { type: string }
        approval: { $ref: '#/components/schemas/ApprovalReceipt' }
      description: Server-authoritative v0.14 request; executor loads normalized intent, snapshot, and decision from store before compiling a plan.
    ExecutionPlanSummary:
      type: object
      additionalProperties: false
      required: [execution_id, account_id, normalized_intent_id, snapshot_id, snapshot_hash, decision_id, decision_hash, approval_id, approval_hash, plan_hash, status, condition_id, token_id, side, quantity_bound, limit_price, time_in_force, collateral_profile_id, max_exposure, executor_version, contract_version, explanation]
      properties:
        execution_id: { type: string }
        account_id: { type: string }
        normalized_intent_id: { type: string }
        snapshot_id: { type: string }
        snapshot_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        decision_id: { type: string }
        decision_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        approval_id: { type: string }
        approval_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        plan_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        status: { type: string, enum: [READY, BLOCKED] }
        condition_id: { type: string }
        token_id: { type: string }
        side: { type: string, enum: [BUY, SELL] }
        quantity_bound: { $ref: '#/components/schemas/QuantityBound' }
        limit_price: { $ref: '#/components/schemas/CanonicalDecimal' }
        time_in_force: { type: string, enum: [GTC, FOK, GTD, FAK] }
        collateral_profile_id:
          oneOf:
            - { type: string }
            - { type: 'null' }
        max_exposure: { $ref: '#/components/schemas/CanonicalDecimal' }
        executor_version: { type: string }
        contract_version: { type: string }
        explanation:
          type: array
          items: { type: string }
    SubmitRequest:
      type: object
      additionalProperties: false
      required: [execution_id, plan_hash, idempotency_key, mode]
      properties:
        execution_id: { type: string }
        plan_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        idempotency_key: { type: string }
        mode: { type: string, enum: [BLOCKED_DRY_RUN, LIVE] }
      description: Idempotency identity is executor-resolved as account_id + execution_id + idempotency_key; submit_attempt is generated internally.
    SubmitReceipt:
      type: object
      additionalProperties: false
      required: [execution_id, receipt_id, status, executor_version, contract_version]
      properties:
        execution_id: { type: string }
        receipt_id: { type: string }
        status: { type: string, enum: [ACCEPTED, POSTED, PARTIAL_REMOTE_UNKNOWN, REMOTE_UNKNOWN, REJECTED, BLOCKED] }
        executor_version: { type: string }
        contract_version: { type: string }
    SignOnlyLifecycleRecord:
      type: object
      additionalProperties: false
      required: [execution_id, account_id, state, event, signed_order_ref, no_remote_side_effect]
      properties:
        event_id: { type: integer, readOnly: true }
        created_at: { type: string, format: date-time, readOnly: true }
        execution_id: { type: string }
        account_id: { type: string }
        state:
          type: string
          enum: [PLANNED, RESERVATION_PREPARED, SIGNING_REQUESTED, SIGNED_DRY_RUN, FAILED, ABANDONED]
        event:
          type: string
          enum: [PREPARE_RESERVATION, REQUEST_SIGNING, SIGNED_WITHOUT_POST, SIGNING_FAILED, ABANDON]
        client_event_id: { type: [string, 'null'], description: Optional execution-scoped idempotency key for append/replay }
        signed_order_ref: { type: [string, 'null'] }
        no_remote_side_effect: { type: boolean }
    StandardSignOnlyConstructionRequest:
      type: object
      additionalProperties: false
      required: [execution_id, account_id, plan_hash, no_remote_side_effect]
      properties:
        execution_id: { type: string }
        account_id: { type: string }
        plan_hash: { type: string }
        signed_order_ref:
          type: [string, 'null']
          pattern: '^sign-only:'
          description: Optional redacted reference to executor-owned sign-only construction output; omitted requests derive an executor-owned sign-only reference and never expose a raw signed order.
        signed_order_digest:
          type: [string, 'null']
          pattern: '^[0-9A-Fa-f]{64}$'
          description: Optional SHA-256 digest binding for the redacted construction artifact; omitted requests derive a deterministic digest from server-authoritative plan metadata.
        no_remote_side_effect: { type: boolean, enum: [true] }
    StandardSignOnlyConstructionReceipt:
      type: object
      additionalProperties: false
      required: [execution_id, signed_order_ref, signed_order_digest, lifecycle_records, no_remote_side_effect]
      properties:
        execution_id: { type: string }
        signed_order_ref:
          type: string
          pattern: '^sign-only:'
        signed_order_digest:
          type: [string, 'null']
          pattern: '^[0-9A-Fa-f]{64}$'
        lifecycle_records:
          type: array
          items: { $ref: '#/components/schemas/SignOnlyLifecycleRecord' }
        no_remote_side_effect: { type: boolean, enum: [true] }
...
    ExecutionLifecycleEvent:
      type: object
      additionalProperties: false
      required: [execution_id, account_id, event_type, event_source, payload]
      properties:
        event_id: { type: integer, readOnly: true }
        created_at: { type: string, format: date-time, readOnly: true }
        execution_id: { type: string }
        account_id: { type: string }
        event_type: { type: string }
        event_source: { type: string }
        payload: { $ref: '#/components/schemas/RedactedPayloadEnvelope' }
    OrderLifecycleEventRecord:
      type: object
      additionalProperties: false
      required: [order_id, event, event_source, payload]
      properties:
        event_id: { type: [integer, 'null'], readOnly: true }
        order_id: { type: string }
        event: { $ref: '#/components/schemas/OrderEventKind' }
        event_source: { type: string }
        correlation_id: { type: [string, 'null'] }
        payload:
          type: object
          additionalProperties: true
        created_at: { type: [string, 'null'], format: date-time, readOnly: true }
    AdminSession:
      type: object
      additionalProperties: false
      required: [principal_subject, scopes, capabilities, no_remote_side_effect]
      properties:
        principal_subject: { type: string, minLength: 1 }
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [ADMIN, ADMIN_READ, ADMIN_CANCEL, EMERGENCY_OPERATOR]
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [READ_AUDIT, CANCEL_ORDER, CANCEL_MARKET, RECONCILE, KILL_SWITCH]
        no_remote_side_effect: { type: boolean, const: true }
    AdminAuditEvent:
      type: object
      additionalProperties: false
      required: [principal_subject, operation, request_fingerprint, result]
      properties:
        audit_id: { type: integer, readOnly: true }
        created_at: { type: string, format: date-time, readOnly: true }
        principal_subject: { type: string }
        operation: { type: string }
        request_fingerprint: { type: [string, 'null'] }
        correlation_id: { type: [string, 'null'] }
        result: { type: string }
...
    RuntimeWorkerStatusReport:
      type: object
      additionalProperties: false
      required: [heartbeats, observations]
      properties:
        heartbeats:
          type: array
          items: { $ref: '#/components/schemas/RuntimeWorkerHeartbeat' }
        observations:
          type: array
          items: { $ref: '#/components/schemas/RuntimeWorkerObservation' }
...
    PortfolioProjection:
      type: object
      additionalProperties: false
      required: [account_id, fills, positions, open_orders, exposure, observed_at_ms]
      properties:
        account_id: { type: string }
        fills:
          type: array
          items: { $ref: '#/components/schemas/FillRecord' }
        positions:
          type: array
          items: { $ref: '#/components/schemas/PositionProjection' }
        open_orders:
          type: array
          items: { $ref: '#/components/schemas/OpenOrderProjection' }
        exposure: { $ref: '#/components/schemas/ExposureProjection' }
        observed_at_ms: { type: integer }
    PortfolioProjectionRecordResponse:
      type: object
      additionalProperties: false
      required: [account_id, observed_at_ms, no_remote_side_effect]
      properties:
        account_id: { type: string }
        observed_at_ms: { type: integer }
        no_remote_side_effect: { type: boolean, const: true }
    RiskLimits:
      type: object
      additionalProperties: false
      required: [max_gross_notional, max_open_order_notional, kill_switch_active]
      properties:
        max_gross_notional: { $ref: '#/components/schemas/CanonicalDecimal' }
        max_open_order_notional: { $ref: '#/components/schemas/CanonicalDecimal' }
        kill_switch_active: { type: boolean }
    RiskDecision:
      oneOf:
        - type: object
          additionalProperties: false
          required: [decision]
          properties:
            decision: { type: string, enum: [ALLOW] }
        - type: object
          additionalProperties: false
          required: [decision, reason]
          properties:
            decision: { type: string, enum: [BLOCK] }
            reason:
              type: string
              enum: [KILL_SWITCH_ACTIVE, GROSS_EXPOSURE_EXCEEDED, OPEN_ORDER_EXPOSURE_EXCEEDED]
    CancelOrderRequest:
      type: object
      additionalProperties: false
      required: [account_id, order_id, reason]
      properties:
        account_id: { type: string }
        order_id: { type: string }
        execution_id: { type: [string, 'null'] }
        reason: { type: string }
    CancelReceipt:
      type: object
      additionalProperties: false
      required: [cancel_id, order_id, state]
      properties:
        cancel_id: { type: string }
        order_id: { type: string }
        state: { type: string, enum: [REQUESTED, REMOTE_ACCEPTED, CONFIRMED_CANCELED, NOT_CANCELED, REMOTE_UNKNOWN, RECONCILE_REQUIRED] }
...
    ReconcileOrderLocalRequest:
      type: object
      additionalProperties: false
      required: [account_id, order_id, remote_observation, reason]
      properties:
        account_id: { type: string }
        order_id: { type: string }
        remote_observation: { $ref: '#/components/schemas/RemoteOrderObservation' }
        reason: { type: string }
    ReconcileOrderLocalResponse:
      type: object
      additionalProperties: false
      required: [order_id, divergence, updated_order, no_remote_side_effect]
      properties:
        order_id: { type: string }
        divergence: { $ref: '#/components/schemas/OrderLifecycleDivergence' }
        updated_order:
          anyOf:
            - $ref: '#/components/schemas/OrderLifecycleRecord'
            - type: 'null'
        no_remote_side_effect: { type: boolean }
    KillSwitchRequest:
      type: object
      additionalProperties: false
      required: [scope, enabled, reason]
      properties:
        scope:
          type: string
          enum: [ACCOUNT, GLOBAL]
        account_id:
          type: [string, 'null']
          description: Required for ACCOUNT scope and omitted for GLOBAL scope.
        enabled: { type: boolean }
        reason: { type: string }
    KillSwitchReceipt:
      type: object
      additionalProperties: false
      required: [scope, enabled, changed_at, effective_at, state_version, persisted, reason]
      properties:
        scope:
          type: string
          enum: [ACCOUNT, GLOBAL]
        account_id: { type: [string, 'null'] }
        enabled: { type: boolean }
        changed_at: { type: string, format: date-time }
        effective_at: { type: string, format: date-time }
        state_version: { type: integer, format: int64 }
        persisted: { type: boolean }
        reason: { type: string }
    ReconcileRequest:
      type: object
      additionalProperties: false
      required: [account_id, execution_id, reason]
      properties:
        account_id: { type: string }
        execution_id: { type: [string, 'null'] }
        order_id: { type: [string, 'null'] }
        remote_observation:
          anyOf:
            - $ref: '#/components/schemas/RemoteOrderObservation'
            - type: 'null'
        reason: { type: string }
    ReconcileReport:
      type: object
      additionalProperties: false
      required: [reconcile_id, status, checked_orders, findings]
      properties:
        reconcile_id: { type: string }
        status: { type: string }
        checked_orders: { type: integer, minimum: 0 }
        findings:
          type: array
          items: { type: string }
...